Moved configs to included files
