+++ addons/factions/cis/CfgVehicles.hpp
@@ -1,684 +1,5 @@
 class CfgVehicles {
-    class CLASS(OPFOR_Unit_Base);
-    class CLASS(CIS_Unit_Base): CLASS(OPFOR_Unit_Base) {
-        SCOPE_HIDDEN;
-        faction = QFACTION(CIS);
-    };
-    class CLASS(CIS_Unit_Droid_Base): CLASS(CIS_Unit_Base) {
-        model = "\MRC\JLTS\characters\DroidArmor\DroidUniformB1.p3d";
-        hiddenSelections[] = {"camo1"};
-
-        armor = 2;
-        armorStructural = 3;
-        explosionShielding = 0.3;
-        impactDamageMultiplier = 0.5;
-        minTotalDamageThreshold = 0.001;
-
-        identityTypes[] = {"lsd_voice_b1Droid"};
-
-        impactEffectsBlood = "ImpactMetal";
-        impactEffectsNoBlood = "ImpactMetal";
-
-        weapons[] = {
-            QCLASS(E5),
-            "Throw",
-            "Put"
-        };
-        respawnWeapons[] = {
-            QCLASS(E5),
-            "Throw",
-            "Put"
-        };
-        magazines[] = {
-            ITEM_5(QCLASS(Mag_100rnd_E5)),
-            "ls_mag_classC_thermalDet"
-        };
-        respawnMagazines[] = {
-            ITEM_5(QCLASS(Mag_100rnd_E5)),
-            "ls_mag_classC_thermalDet"
-        };
-
-        linkedItems[] = {
-            "JLTS_NVG_droid_chip_1", "SWLB_comlink_droid", LINKED_ITEMS
-        };
-        respawnLinkedItems[] = {
-            "JLTS_NVG_droid_chip_1", "SWLB_comlink_droid", LINKED_ITEMS
-        };
-
-        class SoundBleeding {breath[] = {};};
-        class SoundBreath {breath[] = {};};
-        class SoundBurning {breath[] = {};};
-        class SoundChoke {breath[] = {};};
-        class SoundDrown {breath[] = {};};
-        class SoundEnvironExt {
-            generic[] = {
-                {"walk", {QPATHTOF(cis\data\audio\walk\Step1.wav), 2, 1, 15}},
-                {"walk", {QPATHTOF(cis\data\audio\walk\Step2.wav), 2, 1, 15}},
-                {"walk", {QPATHTOF(cis\data\audio\walk\Step3.wav), 2, 1, 15}},
-                {"walk", {QPATHTOF(cis\data\audio\walk\Step4.wav), 2, 1, 15}},
-                {"tactical", {QPATHTOF(cis\data\audio\walk\Step1.wav), 2, 1, 15}},
-                {"tactical", {QPATHTOF(cis\data\audio\walk\Step2.wav), 2, 1, 15}},
-                {"tactical", {QPATHTOF(cis\data\audio\walk\Step3.wav), 2, 1, 15}},
-                {"tactical", {QPATHTOF(cis\data\audio\walk\Step4.wav), 2, 1, 15}},
-                {"run", {QPATHTOF(cis\data\audio\walk\Step1.wav), 2, 1, 30}},
-                {"run", {QPATHTOF(cis\data\audio\walk\Step2.wav), 2, 1, 30}},
-                {"run", {QPATHTOF(cis\data\audio\walk\Step3.wav), 2, 1, 30}},
-                {"run", {QPATHTOF(cis\data\audio\walk\Step4.wav), 2, 1, 30}},
-                {"sprint", {QPATHTOF(cis\data\audio\walk\Step1.wav), 2, 1, 45}},
-                {"sprint", {QPATHTOF(cis\data\audio\walk\Step2.wav), 2, 1, 45}},
-                {"sprint", {QPATHTOF(cis\data\audio\walk\Step3.wav), 2, 1, 45}},
-                {"sprint", {QPATHTOF(cis\data\audio\walk\Step4.wav), 2, 1, 45}}
-            };
-        };
-        class SoundEquipment {
-            soldier[] = {
-                {"walk", {QPATHTOF(cis\data\audio\walk\Step1.wav), 2, 1, 15}},
-                {"walk", {QPATHTOF(cis\data\audio\walk\Step2.wav), 2, 1, 15}},
-                {"walk", {QPATHTOF(cis\data\audio\walk\Step3.wav), 2, 1, 15}},
-                {"walk", {QPATHTOF(cis\data\audio\walk\Step4.wav), 2, 1, 15}},
-                {"tactical", {QPATHTOF(cis\data\audio\walk\Step1.wav), 2, 1, 15}},
-                {"tactical", {QPATHTOF(cis\data\audio\walk\Step2.wav), 2, 1, 15}},
-                {"tactical", {QPATHTOF(cis\data\audio\walk\Step3.wav), 2, 1, 15}},
-                {"tactical", {QPATHTOF(cis\data\audio\walk\Step4.wav), 2, 1, 15}},
-                {"run", {QPATHTOF(cis\data\audio\walk\Step1.wav), 2, 1, 30}},
-                {"run", {QPATHTOF(cis\data\audio\walk\Step2.wav), 2, 1, 30}},
-                {"run", {QPATHTOF(cis\data\audio\walk\Step3.wav), 2, 1, 30}},
-                {"run", {QPATHTOF(cis\data\audio\walk\Step4.wav), 2, 1, 30}},
-                {"sprint", {QPATHTOF(cis\data\audio\walk\Step1.wav), 2, 1, 45}},
-                {"sprint", {QPATHTOF(cis\data\audio\walk\Step2.wav), 2, 1, 45}},
-                {"sprint", {QPATHTOF(cis\data\audio\walk\Step3.wav), 2, 1, 45}},
-                {"sprint", {QPATHTOF(cis\data\audio\walk\Step4.wav), 2, 1, 45}}
-            };
-        };
-        class SoundInjured {breath[] = {};};
-        class SoundRecovered {breath[] = {};};
-    };
-
-    class CLASS(CIS_Unit_Droid_B1): CLASS(CIS_Unit_Droid_Base) {
-        SCOPE_PUBLIC;
-        displayName = "B1 Battle Droid";
-        uniformClass = QCLASS(CIS_Uniform_Droid_B1);
-
-        editorSubcategory = QEDSUBCAT(CIS_B1s);
-        editorPreview = EDITOR_PREVIEW(CIS_Unit_Droid_B1);
-        hiddenSelectionsTextures[] = {"\MRC\JLTS\characters\DroidArmor\data\b1_co.paa"};
-        backpack = QCLASS(CIS_Backpack_Droid_B1);
-    };
-
-    class CLASS(CIS_Unit_Droid_B1_Heavy): CLASS(CIS_Unit_Droid_B1) {
-        displayName = "B1 Heavy Droid";
-        icon = "JLTS_iconManSupportGunner";
-
-        editorPreview = EDITOR_PREVIEW(CIS_Unit_Droid_B1_Heavy);
-
-        weapons[] = {
-            QCLASS(E5C_Stock),
-            "Throw",
-            "Put"
-        };
-        respawnWeapons[] = {
-            QCLASS(E5C_Stock),
-            "Throw",
-            "Put"
-        };
-
-        magazines[] = {
-            ITEM_5(QCLASS(Mag_150rnd_E5C)),
-            "ls_mag_classC_thermalDet"
-        };
-        respawnMagazines[] = {
-            ITEM_5(QCLASS(Mag_150rnd_E5C)),
-            "ls_mag_classC_thermalDet"
-        };
-    };
-
-    class CLASS(CIS_Unit_Droid_B1_AT): CLASS(CIS_Unit_Droid_B1) {
-        displayName = "B1 Battle Droid (AT)";
-        icon = "iconManAT";
-
-        editorPreview = EDITOR_PREVIEW(CIS_Unit_Droid_B1_AT);
-        backpack = QCLASS(CIS_Backpack_Droid_B1_predef_AT);
-
-        weapons[] = {
-            QCLASS(E5),
-            QCLASS(E60R_AT),
-            "Throw",
-            "Put"
-        };
-        respawnWeapons[] = {
-            QCLASS(E5),
-            QCLASS(E60R_AT),
-            "Throw",
-            "Put"
-        };
-
-        magazines[] = {
-            ITEM_5(QCLASS(Mag_100rnd_E5)),
-            QCLASS(Mag_1rnd_E60R_AT_AI),
-            "ls_mag_classC_thermalDet"
-        };
-        respawnMagazines[] = {
-            ITEM_5(QCLASS(Mag_100rnd_E5)),
-            QCLASS(Mag_1rnd_E60R_AT_AI),
-            "ls_mag_classC_thermalDet"
-        };
-    };
-
-    class CLASS(CIS_Unit_Droid_B1_AA): CLASS(CIS_Unit_Droid_B1) {
-        displayName = "B1 Battle Droid (AA)";
-        icon = "iconManAT";
-
-        editorPreview = EDITOR_PREVIEW(CIS_Unit_Droid_B1_AA);
-        backpack = QCLASS(CIS_Backpack_Droid_B1_predef_AA);
-
-        weapons[] = {
-            QCLASS(E5),
-            QCLASS(E60R_AA),
-            "Throw",
-            "Put"
-        };
-        respawnWeapons[] = {
-            QCLASS(E5),
-            QCLASS(E60R_AA),
-            "Throw",
-            "Put"
-        };
-
-        magazines[] = {
-            ITEM_5(QCLASS(Mag_100rnd_E5)),
-            QCLASS(Mag_1rnd_E60R_AA_AI),
-            "ls_mag_classC_thermalDet"
-        };
-        respawnMagazines[] = {
-            ITEM_5(QCLASS(Mag_100rnd_E5)),
-            QCLASS(Mag_1rnd_E60R_AA_AI),
-            "ls_mag_classC_thermalDet"
-        };
-    };
-
-    class CLASS(CIS_Unit_Droid_B1_Sniper): CLASS(CIS_Unit_Droid_B1) {
-        displayName = "B1 Sniper Droid";
-        icon = "JLTS_iconManSniper";
-
-        editorPreview = EDITOR_PREVIEW(CIS_Unit_Droid_B1_Sniper);
-
-        weapons[] = {
-            QCLASS(E5S),
-            "Throw",
-            "Put"
-        };
-        respawnWeapons[] = {
-            QCLASS(E5S),
-            "Throw",
-            "Put"
-        };
-
-        magazines[] = {
-            ITEM_10(QCLASS(Mag_20rnd_E5S)),
-            "ls_mag_classC_thermalDet"
-        };
-        respawnMagazines[] = {
-            ITEM_10(QCLASS(Mag_20rnd_E5S)),
-            "ls_mag_classC_thermalDet"
-        };
-    };
-
-    class CLASS(CIS_Unit_Droid_B1_Marine): CLASS(CIS_Unit_Droid_B1) {
-        displayName = "B1 Marine Droid";
-        uniformClass = QCLASS(CIS_Uniform_Droid_B1_Marine);
-
-        editorPreview = EDITOR_PREVIEW(CIS_Unit_Droid_B1_Marine);
-        hiddenSelectionsTextures[] = {"\MRC\JLTS\characters\DroidArmor\data\b1_marine_co.paa"};
-        backpack = QCLASS(CIS_Backpack_Droid_B1_predef_HE);
-
-        weapons[] = {
-            QCLASS(E5),
-            QCLASS(E60R_AT),
-            "Throw",
-            "Put"
-        };
-        respawnWeapons[] = {
-            QCLASS(E5),
-            QCLASS(E60R_AT),
-            "Throw",
-            "Put"
-        };
-
-        magazines[] = {
-            ITEM_5(QCLASS(Mag_100rnd_E5)),
-            QCLASS(Mag_1rnd_E60R_HE_AI),
-            "ls_mag_classC_thermalDet"
-        };
-        respawnMagazines[] = {
-            ITEM_5(QCLASS(Mag_100rnd_E5)),
-            QCLASS(Mag_1rnd_E60R_HE_AI),
-            "ls_mag_classC_thermalDet"
-        };
-    };
-
-    class CLASS(CIS_Unit_Droid_B1_CQC): CLASS(CIS_Unit_Droid_B1) {
-        displayName = "B1 Shotgun Droid";
-        editorPreview = EDITOR_PREVIEW(CIS_Unit_Droid_B1_CQC);
-
-        weapons[] = {
-            QCLASS(SBB3),
-            "Throw",
-            "Put"
-        };
-        respawnWeapons[] = {
-            QCLASS(SBB3),
-            "Throw",
-            "Put"
-        };
-
-        magazines[] = {
-            ITEM_5(QCLASS(Mag_25rnd_SBB3)),
-            "ls_mag_classC_thermalDet"
-        };
-        respawnMagazines[] = {
-            ITEM_5(QCLASS(Mag_25rnd_SBB3)),
-            "ls_mag_classC_thermalDet"
-        };
-    };
-
-    class CLASS(CIS_Unit_Droid_B1_Security): CLASS(CIS_Unit_Droid_B1) {
-        displayName = "B1 Security Droid";
-        uniformClass = QCLASS(CIS_Uniform_Droid_B1_Security);
-
-        editorPreview = EDITOR_PREVIEW(CIS_Unit_Droid_B1_Security);
-        hiddenSelectionsTextures[] = {"\MRC\JLTS\characters\DroidArmor\data\b1_security_co.paa"};
-        backpack = "";
-    };
-
-    class CLASS(CIS_Unit_Droid_B1_Commander): CLASS(CIS_Unit_Droid_B1) {
-        displayName = "B1 Commander Droid";
-        uniformClass = QCLASS(CIS_Uniform_Droid_B1_Commander);
-        icon = "iconManOfficer";
-
-        editorPreview = EDITOR_PREVIEW(CIS_Unit_Droid_B1_Commander);
-        hiddenSelectionsTextures[] = {"\MRC\JLTS\characters\DroidArmor\data\b1_commander_co.paa"};
-        backpack = QCLASS(CIS_Backpack_Droid_B1_Antenna);
-    };
-
-    class CLASS(CIS_Unit_Droid_B1_Crew): CLASS(CIS_Unit_Droid_B1) {
-        displayName = "B1 Crew Droid";
-        uniformClass = QCLASS(CIS_Uniform_Droid_B1_Crew);
-
-        editorPreview = EDITOR_PREVIEW(CIS_Unit_Droid_B1_Crew);
-        hiddenSelectionsTextures[] = {"\MRC\JLTS\characters\DroidArmor\data\b1_crew_co.paa"};
-        backpack = QCLASS(CIS_Backpack_Droid_B1);
-    };
-
-    class CLASS(CIS_Unit_Droid_B1_Pilot): CLASS(CIS_Unit_Droid_B1) {
-        displayName = "B1 Pilot Droid";
-        uniformClass = QCLASS(CIS_Uniform_Droid_B1_Pilot);
-
-        editorPreview = EDITOR_PREVIEW(CIS_Unit_Droid_B1_Pilot);
-        hiddenSelectionsTextures[] = {"\MRC\JLTS\characters\DroidArmor\data\b1_pilot_co.paa"};
-        backpack = "";
-    };
-
-    class CLASS(CIS_Unit_Droid_B1_Saboteur): CLASS(CIS_Unit_Droid_B1) {
-        displayName = "B1 Saboteur Droid";
-        uniformClass = QCLASS(CIS_Uniform_Droid_B1_Saboteur);
-
-        editorSubcategory = QEDSUBCAT(Special);
-        editorPreview = EDITOR_PREVIEW(CIS_Unit_Droid_B1_Rocket);
-        hiddenSelectionsTextures[] = {"\MRC\JLTS\characters\DroidArmor\data\b1_rocket_co.paa"};
-        backpack = QCLASS(CIS_Backpack_Droid_B1_Saboteur);
-    };
-
-    class CLASS(CIS_Unit_Droid_B1_Prototype): CLASS(CIS_Unit_Droid_B1) {
-        displayName = "B1 Prototype Droid";
-        editorPreview = EDITOR_PREVIEW(CIS_Unit_Droid_B1_Prototype);
-
-        uniformClass = QCLASS(CIS_Uniform_Droid_B1_Prototype);
-        hiddenSelectionsTextures[] = {"\MRC\JLTS\characters\DroidArmor\data\b1_prototype_co.paa"};
-
-        weapons[] = {
-            QCLASS(E5),
-            "Throw",
-            "Put"
-        };
-        respawnWeapons[] = {
-            QCLASS(E5),
-            "Throw",
-            "Put"
-        };
-
-        magazines[] = {
-            ITEM_5(QCLASS(Mag_100rnd_E5)),
-            ITEM_2("ls_mag_classC_thermalDet"),
-            ITEM_2("3AS_SmokeWhite")
-        };
-        respawnMagazines[] = {
-            ITEM_5(QCLASS(Mag_100rnd_E5)),
-            ITEM_2("ls_mag_classC_thermalDet"),
-            ITEM_2("3AS_SmokeWhite")
-        };
-    };
-
-    class CLASS(CIS_Unit_Droid_B1_Geonosis): CLASS(CIS_Unit_Droid_B1) {
-        uniformClass = QCLASS(CIS_Uniform_Droid_B1_Geonosis);
-        editorSubcategory = QEDSUBCAT(CIS_B1s_Geonosis);
-        editorPreview = EDITOR_PREVIEW(CIS_Unit_Droid_B1_Geonosis);
-        hiddenSelectionsTextures[] = {"\MRC\JLTS\characters\DroidUnits\data\Droid_B1Geonosis_co.paa"};
-        backpack = QCLASS(CIS_Backpack_Droid_B1_Geonosis);
-    };
-
-    class CLASS(CIS_Unit_Droid_B1_Heavy_Geonosis): CLASS(CIS_Unit_Droid_B1_Heavy) {
-        uniformClass = QCLASS(CIS_Uniform_Droid_B1_Geonosis);
-        editorSubcategory = QEDSUBCAT(CIS_B1s_Geonosis);
-        editorPreview = EDITOR_PREVIEW(CIS_Unit_Droid_B1_Heavy_Geonosis);
-        hiddenSelectionsTextures[] = {"\MRC\JLTS\characters\DroidUnits\data\Droid_B1Geonosis_co.paa"};
-        backpack = QCLASS(CIS_Backpack_Droid_B1_Geonosis);
-    };
-
-    class CLASS(CIS_Unit_Droid_B1_AT_Geonosis): CLASS(CIS_Unit_Droid_B1_AT) {
-        uniformClass = QCLASS(CIS_Uniform_Droid_B1_Geonosis);
-        editorSubcategory = QEDSUBCAT(CIS_B1s_Geonosis);
-        editorPreview = EDITOR_PREVIEW(CIS_Unit_Droid_B1_AT_Geonosis);
-        hiddenSelectionsTextures[] = {"\MRC\JLTS\characters\DroidUnits\data\Droid_B1Geonosis_co.paa"};
-        backpack = QCLASS(CIS_Backpack_Droid_B1_Geonosis_predef_AT);
-    };
-
-    class CLASS(CIS_Unit_Droid_B1_AA_Geonosis): CLASS(CIS_Unit_Droid_B1_AA) {
-        uniformClass = QCLASS(CIS_Uniform_Droid_B1_Geonosis);
-        editorSubcategory = QEDSUBCAT(CIS_B1s_Geonosis);
-        editorPreview = EDITOR_PREVIEW(CIS_Unit_Droid_B1_AA_Geonosis);
-        hiddenSelectionsTextures[] = {"\MRC\JLTS\characters\DroidUnits\data\Droid_B1Geonosis_co.paa"};
-        backpack = QCLASS(CIS_Backpack_Droid_B1_Geonosis_predef_AA);
-    };
-
-    class CLASS(CIS_Unit_Droid_B1_Sniper_Geonosis): CLASS(CIS_Unit_Droid_B1_Sniper) {
-        uniformClass = QCLASS(CIS_Uniform_Droid_B1_Geonosis);
-        editorSubcategory = QEDSUBCAT(CIS_B1s_Geonosis);
-        editorPreview = EDITOR_PREVIEW(CIS_Unit_Droid_B1_Sniper_Geonosis);
-        hiddenSelectionsTextures[] = {"\MRC\JLTS\characters\DroidUnits\data\Droid_B1Geonosis_co.paa"};
-        backpack = QCLASS(CIS_Backpack_Droid_B1_Geonosis);
-    };
-
-    class CLASS(CIS_Unit_Droid_B1_CQC_Geonosis): CLASS(CIS_Unit_Droid_B1_CQC) {
-        uniformClass = QCLASS(CIS_Uniform_Droid_B1_Geonosis);
-        editorSubcategory = QEDSUBCAT(CIS_B1s_Geonosis);
-        editorPreview = EDITOR_PREVIEW(CIS_Unit_Droid_B1_CQC_Geonosis);
-        hiddenSelectionsTextures[] = {"\MRC\JLTS\characters\DroidUnits\data\Droid_B1Geonosis_co.paa"};
-        backpack = QCLASS(CIS_Backpack_Droid_B1_Geonosis);
-    };
-
-    class CLASS(CIS_Unit_Droid_B1_Commander_Geonosis): CLASS(CIS_Unit_Droid_B1_Commander) {
-        uniformClass = QCLASS(CIS_Uniform_Droid_B1_Commander_Geonosis);
-        editorSubcategory = QEDSUBCAT(CIS_B1s_Geonosis);
-        editorPreview = EDITOR_PREVIEW(CIS_Unit_Droid_B1_Commander_Geonosis);
-        hiddenSelectionsTextures[] = {"\MRC\JLTS\characters\DroidUnits\data\Droid_B1Geonosis_co.paa"};
-        backpack = QCLASS(CIS_Backpack_Droid_B1_Antenna_Geonosis);
-    };
-
-    class CLASS(CIS_Unit_Droid_B1_Training): CLASS(CIS_Unit_Droid_B1) {
-        uniformClass = QCLASS(CIS_Uniform_Droid_B1_Training);
-        editorSubcategory = QEDSUBCAT(CIS_B1s_Training);
-        editorPreview = EDITOR_PREVIEW(CIS_Unit_Droid_B1_Training);
-        hiddenSelectionsTextures[] = {"\MRC\JLTS\characters\DroidUnits\data\Droid_B1Training_co.paa"};
-        backpack = QCLASS(CIS_Backpack_Droid_B1_Training);
-    };
-
-    class CLASS(CIS_Unit_Droid_B1_Heavy_Training): CLASS(CIS_Unit_Droid_B1_Heavy) {
-        uniformClass = QCLASS(CIS_Uniform_Droid_B1_Training);
-        editorSubcategory = QEDSUBCAT(CIS_B1s_Training);
-        editorPreview = EDITOR_PREVIEW(CIS_Unit_Droid_B1_Heavy_Training);
-        hiddenSelectionsTextures[] = {"\MRC\JLTS\characters\DroidUnits\data\Droid_B1Training_co.paa"};
-        backpack = QCLASS(CIS_Backpack_Droid_B1_Training);
-    };
-
-    class CLASS(CIS_Unit_Droid_B1_AT_Training): CLASS(CIS_Unit_Droid_B1_AT) {
-        uniformClass = QCLASS(CIS_Uniform_Droid_B1_Training);
-        editorSubcategory = QEDSUBCAT(CIS_B1s_Training);
-        editorPreview = EDITOR_PREVIEW(CIS_Unit_Droid_B1_AT_Training);
-        hiddenSelectionsTextures[] = {"\MRC\JLTS\characters\DroidUnits\data\Droid_B1Training_co.paa"};
-        backpack = QCLASS(CIS_Backpack_Droid_B1_Training_predef_AT);
-    };
-
-    class CLASS(CIS_Unit_Droid_B1_AA_Training): CLASS(CIS_Unit_Droid_B1_AA) {
-        uniformClass = QCLASS(CIS_Uniform_Droid_B1_Training);
-        editorSubcategory = QEDSUBCAT(CIS_B1s_Training);
-        editorPreview = EDITOR_PREVIEW(CIS_Unit_Droid_B1_AA_Training);
-        hiddenSelectionsTextures[] = {"\MRC\JLTS\characters\DroidUnits\data\Droid_B1Training_co.paa"};
-        backpack = QCLASS(CIS_Backpack_Droid_B1_Training_predef_AA);
-    };
-
-    class CLASS(CIS_Unit_Droid_B1_Sniper_Training): CLASS(CIS_Unit_Droid_B1_Sniper) {
-        uniformClass = QCLASS(CIS_Uniform_Droid_B1_Training);
-        editorSubcategory = QEDSUBCAT(CIS_B1s_Training);
-        editorPreview = EDITOR_PREVIEW(CIS_Unit_Droid_B1_Sniper_Training);
-        hiddenSelectionsTextures[] = {"\MRC\JLTS\characters\DroidUnits\data\Droid_B1Training_co.paa"};
-        backpack = QCLASS(CIS_Backpack_Droid_B1_Training);
-    };
-
-    class CLASS(CIS_Unit_Droid_B1_CQC_Training): CLASS(CIS_Unit_Droid_B1_CQC) {
-        uniformClass = QCLASS(CIS_Uniform_Droid_B1_Training);
-        editorSubcategory = QEDSUBCAT(CIS_B1s_Training);
-        editorPreview = EDITOR_PREVIEW(CIS_Unit_Droid_B1_CQC_Training);
-        hiddenSelectionsTextures[] = {"\MRC\JLTS\characters\DroidUnits\data\Droid_B1Training_co.paa"};
-        backpack = QCLASS(CIS_Backpack_Droid_B1_Training);
-    };
-
-    class CLASS(CIS_Unit_Droid_B1_Commander_Training): CLASS(CIS_Unit_Droid_B1) {
-        uniformClass = QCLASS(CIS_Uniform_Droid_B1_Training);
-        editorSubcategory = QEDSUBCAT(CIS_B1s_Training);
-        editorPreview = EDITOR_PREVIEW(CIS_Unit_Droid_B1_Commander_Training);
-        hiddenSelectionsTextures[] = {"\MRC\JLTS\characters\DroidUnits\data\Droid_B1Training_co.paa"};
-        backpack = QCLASS(CIS_Backpack_Droid_B1_Training);
-    };
-
-    class CLASS(CIS_Unit_Droid_BX): CLASS(CIS_Unit_Droid_Base) {
-        SCOPE_PUBLIC;
-
-        editorSubcategory = QEDSUBCAT(SpecialForces);
-        editorPreview = EDITOR_PREVIEW(CIS_Unit_Droid_BX);
-
-        displayName = "BX Commando Droid";
-        uniformClass = QCLASS(CIS_Uniform_Droid_BX);
-
-        model = "\ls_armor_redfor\uniform\cis\bx\lsd_cis_bx_uniform.p3d";
-        hiddenSelectionsTextures[] = {"\ls_armor_redfor\uniform\cis\bx\data\body_co.paa"};
-
-        magazines[] = {
-            ITEM_5(QCLASS(Mag_100rnd_E5)),
-            ITEM_2("ls_mag_classC_thermalDet"),
-            ITEM_3("3AS_SmokeWhite")
-        };
-        respawnMagazines[] = {
-            ITEM_5(QCLASS(Mag_100rnd_E5)),
-            ITEM_2("ls_mag_classC_thermalDet"),
-            ITEM_3("3AS_SmokeWhite")
-        };
-        items[] = {
-            // Other
-            ITEM_5("ACE_CableTie")
-        };
-        respawnItems[] = {
-            // Other
-            ITEM_5("ACE_CableTie")
-        };
-
-        linkedItems[] = {
-            QCLASS(CIS_Vest_Droid_BX), "JLTS_NVG_droid_chip_2", "SWLB_comlink_droid", LINKED_ITEMS
-        };
-        respawnLinkedItems[] = {
-            QCLASS(CIS_Vest_Droid_BX), "JLTS_NVG_droid_chip_2", "SWLB_comlink_droid", LINKED_ITEMS
-        };
-    };
-
-    class CLASS(Other_Backpack_Base);
-    class CLASS(CIS_Backpack_Droid_B1): CLASS(Other_Backpack_Base) {
-        SCOPE_PUBLIC;
-
-        displayName = "[CIS] Battle Droid Backpack";
-
-        model = "\MRC\JLTS\characters\DroidArmor\DroidBackpackB1.p3d";
-        hiddenSelections[] = {"camo1"};
-        hiddenSelectionsTextures[] = {"\MRC\JLTS\characters\DroidArmor\data\b1_backpack_co.paa"};
-        picture = "\MRC\JLTS\characters\DroidArmor\data\ui\b1_backpack_ui_ca.paa";
-
-        tf_dialog = "JLTS_droid_lr_radio_dialog";
-        tf_dialogUpdate = "call TFAR_fnc_updateLRDialogToChannel;";
-        tf_encryptionCode = "tf_east_radio_code";
-        tf_hasLRradio = TRUE;
-        tf_range = 15000;
-        tf_subtype = "digital_lr";
-
-        class XtdGearInfo {
-            model = QCLASS(CIS_Backpacks_B1);
-            camo = "Standard";
-            variant = "Standard";
-        };
-    };
-
-    class CLASS(CIS_Backpack_Droid_B1_predef_HE): CLASS(CIS_Backpack_Droid_B1) {
-        SCOPE_HIDDEN;
-        class TransportMagazines {
-            MAG_XX(CLASS(Mag_1rnd_E60R_HE),3);
-        };
-    };
-
-    class CLASS(CIS_Backpack_Droid_B1_predef_AT): CLASS(CIS_Backpack_Droid_B1) {
-        SCOPE_HIDDEN;
-        class TransportMagazines {
-            MAG_XX(CLASS(Mag_1rnd_E60R_AT),3);
-        };
-    };
-
-    class CLASS(CIS_Backpack_Droid_B1_predef_AA): CLASS(CIS_Backpack_Droid_B1) {
-        SCOPE_HIDDEN;
-        class TransportMagazines {
-            MAG_XX(CLASS(Mag_1rnd_E60R_AA),3);
-        };
-    };
-
-    class CLASS(CIS_Backpack_Droid_B1_Saboteur): CLASS(CIS_Backpack_Droid_B1) {
-        displayName = "[CIS] Battle Droid Backpack (Saboteur)";
-        hiddenSelectionsTextures[] = {QPATHTOF(cis\data\textures\backpacks\B1_Saboteur_co.paa)};
-
-        class XtdGearInfo: XtdGearInfo {
-            camo = "Saboteur";
-        };
-    };
-
-    class CLASS(CIS_Backpack_Droid_B1_Prototype): CLASS(CIS_Backpack_Droid_B1) {
-        displayName = "[CIS] Battle Droid Backpack (Prototype)";
-        hiddenSelectionsTextures[] = {"\MRC\JLTS\characters\DroidArmor\data\b1_backpack_prototype_co.paa"};
-
-        class XtdGearInfo: XtdGearInfo {
-            camo = "Prototype";
-        };
-    };
-
-    class CLASS(CIS_Backpack_Droid_B1_Geonosis): CLASS(CIS_Backpack_Droid_B1) {
-        displayName = "[CIS] Battle Droid Backpack (Geonosis)";
-        hiddenSelectionsTextures[] = {"\MRC\JLTS\characters\DroidUnits\data\Droid_B1Geonosis_backpack_co.paa"};
-
-        class XtdGearInfo: XtdGearInfo {
-            variant = "Geonosis";
-        };
-    };
-
-    class CLASS(CIS_Backpack_Droid_B1_Geonosis_predef_AT): CLASS(CIS_Backpack_Droid_B1_predef_AT) {
-        hiddenSelectionsTextures[] = {"\MRC\JLTS\characters\DroidUnits\data\Droid_B1Geonosis_backpack_co.paa"};
-    };
-
-    class CLASS(CIS_Backpack_Droid_B1_Geonosis_predef_AA): CLASS(CIS_Backpack_Droid_B1_predef_AA) {
-        hiddenSelectionsTextures[] = {"\MRC\JLTS\characters\DroidUnits\data\Droid_B1Geonosis_backpack_co.paa"};
-    };
-
-    class CLASS(CIS_Backpack_Droid_B1_Training): CLASS(CIS_Backpack_Droid_B1) {
-        displayName = "[CIS] Battle Droid Backpack (Training)";
-        hiddenSelectionsTextures[] = {"\MRC\JLTS\characters\DroidUnits\data\Droid_B1Training_backpack_co.paa"};
-
-        class XtdGearInfo: XtdGearInfo {
-            variant = "Training";
-        };
-    };
-
-    class CLASS(CIS_Backpack_Droid_B1_Training_predef_AT): CLASS(CIS_Backpack_Droid_B1_predef_AT) {
-        hiddenSelectionsTextures[] = {"\MRC\JLTS\characters\DroidUnits\data\Droid_B1Training_backpack_co.paa"};
-    };
-
-    class CLASS(CIS_Backpack_Droid_B1_Training_predef_AA): CLASS(CIS_Backpack_Droid_B1_predef_AA) {
-        hiddenSelectionsTextures[] = {"\MRC\JLTS\characters\DroidUnits\data\Droid_B1Training_backpack_co.paa"};
-    };
-
-    class CLASS(CIS_Backpack_Droid_B1_Antenna): CLASS(CIS_Backpack_Droid_B1) {
-        displayName = "[CIS] Battle Droid Antenna";
-        model = "\MRC\JLTS\characters\DroidArmor\DroidAntennaB1.p3d";
-        hiddenSelectionsTextures[] = {"\MRC\JLTS\characters\DroidArmor\data\b1_antenna_co.paa"};
-        picture = "\MRC\JLTS\characters\DroidArmor\data\ui\b1_antenna_ui_ca.paa";
-
-        class XtdGearInfo: XtdGearInfo {
-            camo = "Antenna";
-            variant = "Standard";
-        };
-    };
-
-    class CLASS(CIS_Backpack_Droid_B1_Antenna_Geonosis): CLASS(CIS_Backpack_Droid_B1_Antenna) {
-        displayName = "[CIS] Battle Droid Antenna (Geonosis)";
-        hiddenSelectionsTextures[] = {"\MRC\JLTS\characters\DroidUnits\data\Droid_B1Geonosis_antenna_co.paa"};
-
-        class XtdGearInfo: XtdGearInfo {
-            variant = "Geonosis";
-        };
-    };
-
-    class CLASS(CIS_Backpack_Droid_B1_Antenna_Training): CLASS(CIS_Backpack_Droid_B1_Antenna) {
-        displayName = "[CIS] Battle Droid Antenna (Training)";
-        hiddenSelectionsTextures[] = {"\MRC\JLTS\characters\DroidUnits\data\Droid_B1Training_antenna_co.paa"};
-
-        class XtdGearInfo: XtdGearInfo {
-            variant = "Training";
-        };
-    };
-
-    class CLASS(CIS_Jetpack_Droid): CLASS(CIS_Backpack_Droid_B1) {
-        displayName = "[CIS] Battle Droid Jetpack";
-
-        model = "\MRC\JLTS\characters\DroidArmor\DroidJetpackB1.p3d";
-        hiddenSelections[] = {"camo1"};
-        hiddenSelectionsTextures[] = {"\MRC\JLTS\characters\DroidArmor\data\b1_jetpack_co.paa"};
-        picture = "\MRC\JLTS\characters\DroidArmor\data\ui\b1_jetpack_ui_ca.paa";
-
-        JLTS_isJumppack = TRUE;
-        JLTS_settingsDialog = "JLTS_droid_jumppack_dialog";
-
-        JLTS_energy[] = {25,150};
-        JLTS_jumpLimits[] = {50, {2, 50}, "default"};
-        JLTS_steeringCoef = 0.06;
-        JLTS_recharge = 2;
-
-        JLTS_effectPoints[] = {"effect_left", "effect_right"};
-        JLTS_sounds[] = {
-            {"\MRC\JLTS\jumppacks\sounds\ignite.wss", 1, 1},
-            {"\MRC\JLTS\jumppacks\sounds\ignite_fail.wss", 3, 1},
-            {"\MRC\JLTS\jumppacks\sounds\stop.wss", 1, 1},
-            {"\MRC\JLTS\jumppacks\sounds\stop_emergency.wss", 5, 1},
-            "JLTS_jumppack_error",
-            "JLTS_SFX_jumppack_idle"
-        };
-    };
-
-    class CLASS(CIS_Jetpack_DroidRocket): CLASS(CIS_Jetpack_Droid) {
-        displayName = "[KC] Battle Droid Jetpack (Rocket)";
-        hiddenSelectionsTextures[] = {"\MRC\JLTS\characters\DroidArmor\data\b1_jetpack_rocket_co.paa"};
-    };
-
-    class CLASS(AAT_Base);
-    class CLASS(AAT_CIS): CLASS(AAT_Base) {
-        SCOPE_PUBLIC;
-
-        faction = QFACTION(CIS);
-
-        crew = QCLASS(CIS_Unit_Droid_B1_Crew);
-        typicalCargo[] = {QCLASS(CIS_Unit_Droid_B1_Crew)};
-
-        hiddenSelectionsTextures[] = {"\3AS\3AS_AAT\data\CIS_AAT_CO.paa"};
-        editorPreview = EEDITOR_PREVIEW(vehicles\land\aat,AAT_Blue);
-    };
+    #include "configs\Units.hpp"
+    #include "configs\Backpacks.hpp"
+    #include "configs\Vehicles.hpp"
 };
